+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 
 ## Unreleased
 
+### Fixed
+* [`no-unused-state`]: avoid a crash on a class field gDSFP ([#3236][] @ljharb)
+
+[#3236]: https://github.com/yannickcr/eslint-plugin-react/issues/3236
+
 ## [7.29.3] - 2022.03.03
 
 ### Fixed